// src/app/model/media.ts
export class Media {
  
   mediaTitle: string;
 
   mediaDescription: string;
  
   effects: string[];
  
   tags: string[];
  
   mediaID: string;
  
   userId: number;

   hidden: boolean;
  
   defualtProfile: boolean;
  
   fileType: string;
 
   fileSize: number;
  
  type: string;

  fileName: string;
}